+++ src/array/array_async_readable.rs
@@ -1,6 +1,5 @@
 use std::sync::Arc;
 
-use super::TriviallyTransmutable;
 use futures::{stream::FuturesUnordered, StreamExt};
 use itertools::Itertools;
 
@@ -11,7 +10,6 @@ use crate::{
 };
 
 use super::{
-    array_errors::TransmuteError,
     codec::{
         ArrayCodecTraits, ArrayToBytesCodecTraits, AsyncArrayPartialDecoderTraits,
         AsyncStoragePartialDecoder,
@@ -30,22 +28,8 @@ macro_rules! array_async_retrieve_elements {
             ))
         } else {
             let bytes = $self.$func($($arg)*).await?;
-            if safe_transmute::align::check_alignment::<_, T>(&bytes).is_ok() {
-                // no-copy path
-                let mut bytes = core::mem::ManuallyDrop::new(bytes);
-                Ok(unsafe {
-                    Vec::from_raw_parts(
-                        bytes.as_mut_ptr().cast::<T>(),
-                        bytes.len() / core::mem::size_of::<T>(),
-                        bytes.len(),
-                    )
-                })
-            } else {
-                let elements = safe_transmute::transmute_many_permissive::<T>(&bytes)
-                    .map_err(TransmuteError::from)?
-                    .to_vec();
-                Ok(elements)
-            }
+            let elements = crate::array::transmute_from_bytes_vec::<T>(bytes);
+            Ok(elements)
         }
     };
 }
@@ -147,7 +131,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///  - `chunk_indices` are invalid,
     ///  - there is a codec decoding error, or
     ///  - an underlying store error.
-    pub async fn async_retrieve_chunk_elements<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_chunk_elements<T: bytemuck::Pod + Send + Sync>(
         &self,
         chunk_indices: &[u64],
     ) -> Result<Vec<T>, ArrayError> {
@@ -167,7 +151,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///
     /// # Panics
     /// Will panic if a chunk dimension is larger than `usize::MAX`.
-    pub async fn async_retrieve_chunk_ndarray<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_chunk_ndarray<T: bytemuck::Pod + Send + Sync>(
         &self,
         chunk_indices: &[u64],
     ) -> Result<ndarray::ArrayD<T>, ArrayError> {
@@ -241,7 +225,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///
     /// # Errors
     /// Returns an [`ArrayError`] if the size of `T` does not match the data type size or a [`Array::async_retrieve_chunks`] error condition is met.
-    pub async fn async_retrieve_chunks_elements<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_chunks_elements<T: bytemuck::Pod + Send + Sync>(
         &self,
         chunks: &ArraySubset,
     ) -> Result<Vec<T>, ArrayError> {
@@ -252,7 +236,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///
     /// # Errors
     /// Returns an [`ArrayError`] if the size of `T` does not match the data type size or a [`Array::async_retrieve_chunks`] error condition is met.
-    pub async fn async_retrieve_chunks_ndarray<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_chunks_ndarray<T: bytemuck::Pod + Send + Sync>(
         &self,
         chunks: &ArraySubset,
     ) -> Result<ndarray::ArrayD<T>, ArrayError> {
@@ -485,7 +469,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///  - an array subset is invalid or out of bounds of the array,
     ///  - there is a codec decoding error, or
     ///  - an underlying store error.
-    pub async fn async_retrieve_array_subset_elements<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_array_subset_elements<T: bytemuck::Pod + Send + Sync>(
         &self,
         array_subset: &ArraySubset,
     ) -> Result<Vec<T>, ArrayError> {
@@ -503,7 +487,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///
     /// # Panics
     /// Will panic if any dimension in `chunk_subset` is `usize::MAX` or larger.
-    pub async fn async_retrieve_array_subset_ndarray<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_array_subset_ndarray<T: bytemuck::Pod + Send + Sync>(
         &self,
         array_subset: &ArraySubset,
     ) -> Result<ndarray::ArrayD<T>, ArrayError> {
@@ -574,7 +558,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///  - the chunk subset is invalid,
     ///  - there is a codec decoding error, or
     ///  - an underlying store error.
-    pub async fn async_retrieve_chunk_subset_elements<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_chunk_subset_elements<T: bytemuck::Pod + Send + Sync>(
         &self,
         chunk_indices: &[u64],
         chunk_subset: &ArraySubset,
@@ -597,7 +581,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits> Array<TStorage> {
     ///
     /// # Panics
     /// Will panic if the number of elements in `chunk_subset` is `usize::MAX` or larger.
-    pub async fn async_retrieve_chunk_subset_ndarray<T: TriviallyTransmutable + Send + Sync>(
+    pub async fn async_retrieve_chunk_subset_ndarray<T: bytemuck::Pod + Send + Sync>(
         &self,
         chunk_indices: &[u64],
         chunk_subset: &ArraySubset,