+++ src/array/codec/array_to_bytes/sharding.rs
@@ -185,7 +185,7 @@ mod tests {
         let array_representation =
             ArrayRepresentation::new(chunk_shape, DataType::UInt16, FillValue::from(0u16)).unwrap();
         let elements: Vec<u16> = (0..array_representation.num_elements() as u16).collect();
-        let bytes = safe_transmute::transmute_to_bytes(&elements).to_vec();
+        let bytes = crate::array::transmute_to_bytes_vec(elements);
 
         let codec_configuration: ShardingCodecConfiguration = serde_json::from_str(json).unwrap();
         let codec = ShardingCodec::new_with_configuration(&codec_configuration).unwrap();
@@ -298,7 +298,7 @@ mod tests {
             ArrayRepresentation::new(vec![2, 4, 4], DataType::UInt16, FillValue::from(0u16))
                 .unwrap();
         let elements: Vec<u16> = (0..array_representation.num_elements() as u16).collect();
-        let bytes = safe_transmute::transmute_to_bytes(&elements).to_vec();
+        let bytes = crate::array::transmute_to_bytes_vec(elements);
 
         let codec_configuration: ShardingCodecConfiguration =
             serde_json::from_str(JSON_VALID2).unwrap();